+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning][SemVer].
 ## Unreleased
-
+### Changed
+- Make healthcheck.sh more robust.
 
 ## 0.6.2 - 2023-11-18
 ### Changed